Added Order Quantity Increase and Decrease

+++ store/views.py
@@ -1,3 +1,4 @@
+import imp
 from django.shortcuts import render
 from rest_framework.decorators import api_view
 from store.serializers import CategorySerializer, OrderSerializer, ProductSerializer
@@ -5,6 +6,9 @@ from store.models import Category, Order, Product
 from rest_framework.response import Response
 from django.shortcuts import get_object_or_404
 from rest_framework import status
+from django.contrib import messages
+from rest_framework.views import APIView
+
 
 
 @api_view(['GET','POST'])
@@ -75,28 +79,45 @@ def get_orders(request):
         return Response(serializer.data, status=status.HTTP_201_CREATED)
     return Response(serializer.data,status=status.HTTP_400_BAD_REQUEST)
 
-@api_view(['GET','DELETE'])
-def get_orderitem(request, pk):
-    order=get_object_or_404(Order, product_id=pk)
+
+
+@api_view(['GET','PATCH','DELETE'])
+def get_orderproduct(request, pk):
+    order=Order.objects.get(id=pk)
     if request.method == "GET":
+        order.quantity += 1
+        order.save()
         serializer=OrderSerializer(order)
-        return Response(serializer.data, status=status.HTTP_200_OK)
-    
-    # if request.method == "PUT":
-    #     serializer=OrderSerializer(order, data=request.data)
-    #     serializer.is_valid()
-    #     serializer.save()
-    #     return Response(serializer.data, status=status.HTTP_202_ACCEPTED)
+        return Response(serializer.data, status=status.HTTP_201_CREATED)
+ 
+    # Decrease Order Quantity
+    if request.method == "PATCH":
+        order.quantity -= 1
+        if order.quantity <= 0: #Bug not done
+            return Response(status=status.HTTP_200_OK) #Still Working on it.
+        order.save()
+        serializer=OrderSerializer(order)
+        return Response(serializer.data, status=status.HTTP_202_ACCEPTED)
+
     #Cancel Order
     if request.method =="DELETE":
         order.delete()
-        return Response(serializer.data,status=status.is_success)
-
-
-    
+        serializer=OrderSerializer(order)
+        return Response(serializer.data, status=status.HTTP_200_OK)
 
 
 def add_to_cart(request, pk):
-    product=Product.objects.get(product_id=pk)
-    
+    # Trying to increase Order quantity when being clicked and  calculate total price
+
+
     pass
+    
+# order.quantity += 1
+
+
+# class Checkout(APIView):
+    
+#     def total_price(request, pk):
+#         product=Product.objects.get(product_id=pk)
+#         order = Order.objects.get(user=request.session)
+#         product.price * order.quantity
